Refactored to seperate sections for each widget
