+++ CLAUDE.md
@@ -1,1 +1,5 @@
-@AGENTS.MD
+# CLAUDE.md
+
+This file provides guidance to Claude Code (claude.ai/code) when working with code in this repository.
+
+@AGENTS.md
