+++ interface/input.json
@@ -2,13 +2,13 @@
   "hook_event_name": "Status",
   "session_id": "abc123...",
   "transcript_path": "/Users/jhivan/Developer/play/status-line/test_transcript.json",
-  "cwd": "/Users/jhivan/Developer/play/status-line",
+  "cwd": "/home/jhivandb/Developer/cc-status-line-workspace/status-line-rewrite",
   "model": {
     "id": "claude-opus-4-1",
     "display_name": "Opus"
   },
   "workspace": {
-    "current_dir": "/Users/jhivan/Developer/play/status-line",
+    "current_dir": "/home/jhivandb/Developer/cc-status-line-workspace/status-line-rewrite",
     "project_dir": "/Users/jhivan/Developer/play/status-line"
   },
   "version": "1.0.80",
